+++ docs/harfbuzz-sections.txt
@@ -207,6 +207,7 @@ HB_HAS_GOBJECT
 HB_HAS_GRAPHITE
 HB_HAS_ICU
 HB_HAS_RASTER
+HB_HAS_SUBSET
 HB_HAS_UNISCRIBE
 HB_HAS_VECTOR
 HB_HAS_WASM
