+++ meson.build
@@ -269,6 +269,10 @@ if not get_option('vector').disabled()
   conf.set('HAVE_VECTOR', 1)
 endif
 
+if not get_option('subset').disabled()
+  conf.set('HAVE_SUBSET', 1)
+endif
+
 if freetype_dep.found()
   conf.set('HAVE_FREETYPE', 1)
   check_freetype_funcs = [
