+++ src/hb-raster-svg-base.hh
@@ -218,6 +218,7 @@ hb_raster_svg_tag_is_container (hb_svg_str_t tag)
   return tag.eq ("g") || tag.eq ("a") || tag.eq ("svg") || tag.eq ("symbol");
 }
 
+
 static inline bool
 hb_raster_svg_tag_is_container_or_use (hb_svg_str_t tag)
 {
