+++ src/hb-raster-svg-clip.cc
@@ -89,19 +89,6 @@ svg_clip_append_shape (hb_svg_clip_collect_context_t *ctx,
     *ctx->had_alloc_failure = true;
 }
 
-static void
-svg_skip_subtree (hb_svg_xml_parser_t &parser)
-{
-  int depth = 1;
-  while (depth > 0)
-  {
-    hb_svg_token_type_t tok = parser.next ();
-    if (tok == SVG_TOKEN_EOF) break;
-    if (tok == SVG_TOKEN_CLOSE_TAG) depth--;
-    else if (tok == SVG_TOKEN_OPEN_TAG) depth++;
-  }
-}
-
 static inline bool
 svg_resolve_element_visibility (hb_svg_xml_parser_t &parser,
                                 bool parent_visible)
