+++ src/hb-raster-svg-parse.hh
@@ -202,6 +202,19 @@ struct hb_svg_xml_parser_t
   }
 };
 
+static inline void
+svg_skip_subtree (hb_svg_xml_parser_t &parser)
+{
+  int depth = 1;
+  while (depth > 0)
+  {
+    hb_svg_token_type_t tok = parser.next ();
+    if (tok == SVG_TOKEN_EOF) break;
+    if (tok == SVG_TOKEN_CLOSE_TAG) depth--;
+    else if (tok == SVG_TOKEN_OPEN_TAG) depth++;
+  }
+}
+
 static inline hb_svg_str_t
 svg_pick_attr_or_style (const hb_svg_xml_parser_t &parser,
                         hb_svg_str_t style_value,
